DevOps/gantt_api_v2/requirements.txt
fastapi==0.97.0
PyJWT==1.7.1
httpx==0.23.1
uvicorn==0.22.0
colorlog==6.7.0